requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
geojson==2.5.0
numpy==1.21.0
pandas==1.3.0
pip==21.1.3
pkg-resources==0.0.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
setuptools==44.0.0
six==1.16.0
tornado==6.1
tzlocal==2.1
